backend/backend/requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
djangorestframework==3.12.4
httpie==2.4.0
idna==2.10
jsonfield==3.1.0
Pygments==2.9.0
PyJWT==1.7.1
PySocks==1.7.1
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.1
twilio==6.60.0
typing-extensions==3.10.0.0
urllib3==1.26.5
